requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.2
attrs==23.2.0
auth==0.5.3
bcrypt==4.1.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
eventlet==0.36.1
falcon==3.1.3
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mongoengine==0.28.2
multidict==6.0.5
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
ngrok==1.3.0
oauth2client==4.1.3
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
pyaml==24.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyDrive==1.3.1
PyJWT==2.8.0
pymongo==4.8.0
pyngrok==7.2.0
pyparsing==3.1.2
pypng==0.20220715.0
PyQRCode==1.2.1
python-dotenv==1.0.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.32.3
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4
